Config YAML type (#361)
* Rename Config to PackageJSON

In anticipation of a new SUSHI Config type, rename Config  to PackageJSON.  Also rename corresponding variables from config to packageJSON.  This will help prevent confusions when we are concurrently supporting both.

* New Configuration Type

This configuration type is based on the proposed YAML configuration type, but reflects the data after processing (rather than a straight representation of the YAML format).  I'm checking this in early so people can begin working on exporters.

A future commit will include a YAMLConfiguration type that more closely resembles the actual YAML, as well as an importer to go from one to the other.