Create IGExporter.IG.test.ts, change config parameter default
The tests that exist in IGExporter.test.ts are being skipped for now.
THe new tests that will check for implementation guides are in
IGExporter.IG.test.ts. PackageJSON is still being included so that
the export process is able to finish without throwing an error.

The show-inherited-invariants config parameter is no longer created
by default.