Import config.yaml
Adds support for importing a config.yaml file with very permissive user-friendly types and translating it into the more strict Configuration format.

Handles converting shorthand-style codes to code/Coding/CodeableConcept, shorthand-style quantity syntax to Quantity, optional single-item representation of some array values, checking for required properties, validating code values (where feasible), and handling when YAML parses intended string values as non-strings.

This also does some of the initial hookup into the app.  If a config.yaml is present, it will parse it and put it in the tank and export Package object -- but it's not currently being used in the actual export.