Support all page content folders
Content files for the implementation guide can appear in folders
name pagecontent, pages, and resource-docs.

Parameters for copyrightyear and releaselabel are no longer assigned
default values.

FHIR R4 is no longer skipped as a dependency, because it no longer
needs to be included in the list of dependencies in order to specify
FHIR version.

Properties on ImplementationGuide updated to reflect some changes
to the Configuration type, removing default values and potential
fallback values.

Thank you to SUSHI development team for all of your helpful
review feedback!