Export implementation guide using YAML config (#384)
* Use configuration to build IG JSON

When exporting the JSON file for an implementation guide, the
IGExporter uses the Configuration object, and does not use the
PackageJSON object. Other parts of the IGExporter still use the
PackageJSON object.

Testing support only includes the #simple-ig.

* Create IGExporter.IG.test.ts, change config parameter default

The tests that exist in IGExporter.test.ts are being skipped for now.
THe new tests that will check for implementation guides are in
IGExporter.IG.test.ts. PackageJSON is still being included so that
the export process is able to finish without throwing an error.

The show-inherited-invariants config parameter is no longer created
by default.

* Process YAML page configuration

When pages are configured in YAML, that configuration should be used
to create page attributes in the implementation guide. Page order
and hierarchies are preserved. Files of any type may be listed. An
error is shown if the file for a configured page is not found.

* Avoid including empty page lists

A configured page may have subpages, but because intro and notes
pages are not added to the IG, that page may end up with no subpages
in the IG. In this case, omit the page attribute rather than
including a page attribute that is an empty list.

* Use configuration from package in IGExporter

* Remove commented-out code, add jsdoc for new functions

* Support all page content folders

Content files for the implementation guide can appear in folders
name pagecontent, pages, and resource-docs.

Parameters for copyrightyear and releaselabel are no longer assigned
default values.

FHIR R4 is no longer skipped as a dependency, because it no longer
needs to be included in the list of dependencies in order to specify
FHIR version.

Properties on ImplementationGuide updated to reflect some changes
to the Configuration type, removing default values and potential
fallback values.

Thank you to SUSHI development team for all of your helpful
review feedback!

* Handle groups and resources in configuration

Resources may be assigned to groups based on a groupingId or by
being listed with a configured group.

Resources may be added through configuration to override values in
generated resources or to define entirely new resources.

* Build backup url using configured id

* Add developer note to addResources and addPredefinedResources

Both of these functions deal with adding resources, but in ways
that are just different enough that factoring out common operations
is not really helpful. Comments are added above each function as a
reminder to analyze both functions whenever making any changes
relating to the process of adding resources.

* Add missing properties to IG initialization

Some small changes to tests in the old, skipped IGExporter test
file, but those tests might get reactivated, so keep things tidy
in there.

* Add jurisdiction and copyright to IGExporter

Some day, I will learn to read classes more carefully.

* Re-add test cases for existing resources

These test cases were in IGExporter.test.ts and were being skipped.
They are still necessary, so they are now active again.

* Test for omitting predefined resource

* Don't directly mutate config.dependencies