Rules can be set on children of choice elements (#423)
* Test for applying ContainsRule to child of choice element

* Unfold choice elements

Unfolding of choice elements is based on finding a common ancestor
of the available types. All types are descended from Element, so
there should never be a case where no common ancestor can be found.

* First ancestor is first, not last, element in the intersection

Add test for a choice element whose first ancestor is not Element.

* Add test for invalid path

* Add StructureDefinitions for new test cases

* Use getTypeLineage

The existing getTypeLineage function returns type and profile
metadata, which is used when unfolding choice element types. By
including profile metadata, common ancestors that are profiles will
be found.

* Remove commented-out code