Clean up StructureDefinition.extension during export (#457)
* Clean up StructureDefinition after export

No-path CaretValueRules on a Profile should produce content that
is structurally identical to what would be found on an Instance of
that profile. Extra attributes such as _primitive that are added
during processing should not be present on the exported
StructureDefinition.

* Use cleanResource for all resource cleaning needs

The common function cleanInstance is now cleanResource. It accepts
an optional skip function, which allows it to be used for cleaning
instances and structure definitions.

* Clarify skipFn

* Add test for child of primitive element