Log error when FHIR entity reuses id
The FHIR specification requires that each resource of a given type
has a unique id. Profiles and Extensions are exported as FHIR
StructureDefinitions, so they are checked together for unique ids.
Instances may be many different types, so only log an error if
Instances have the same type and id.