Skip inline instances when checking for duplicated id
Inline instances are not written to their own files. Therefore, it
is acceptable for multiple inline instances of the same type to
have the same id. This will generally be the case, as inline
instance definitions are not automatically assigned an id, so they
would all end up with undefined in their id properties.