Enforce uniqueness for FSH names and FHIR ids (#465)
* Skip and log error when FSH entity reuses name

Because SUSHI tracks FSH entities by name, names should be unique
for each type that is tracked by a map in FSHDocument. If an entity
has a name that belongs to an existing entity of the same type, the
new entity is skipped. This behavior already exists for aliases.

* Log error when FHIR entity reuses id

The FHIR specification requires that each resource of a given type
has a unique id. Profiles and Extensions are exported as FHIR
StructureDefinitions, so they are checked together for unique ids.
Instances may be many different types, so only log an error if
Instances have the same type and id.

* Log error when mappings on Structure Definition share id

Mapping id should be unique for each mapping in a Structure
Definition. Mapping id can be duplicated on different Structure
Definitions.

* Skip inline instances when checking for duplicated id

Inline instances are not written to their own files. Therefore, it
is acceptable for multiple inline instances of the same type to
have the same id. This will generally be the case, as inline
instance definitions are not automatically assigned an id, so they
would all end up with undefined in their id properties.

* Update tests for duplicate FSH entities

* Inline instance and non-inline instance may share id

Inline instances are not FHIR resources, so the id uniqueness rule
does not apply. Skip the id uniqueness check when the current
instance is an inline instance, and do not check against inline
instances already in the package.