Use expected ig-data path even when it doesn't exist on the filesystem (#489)
The IGExporter expects a path to be passed in, even if the path does not exist on the file system. If it receives null, it throws.  Fix app.ts to never pass in a null path to the IGExporter.

Since we only get the path in one place, and since it is now a single line of code, we no longer have need for a separate getIgDataPath function.