Add resources to StructureDefinition
The "contained" element on a StructureDefinition can contain other
Resources. Since this element is on the StructureDefinition itself,
these Resources are added with a CaretValueRule. Since the Resource
to add may be an inline Instance, these rules must be deferred
until after the InstanceExporter has finished. But, any Resource
that can be fished up can be added.