Detect non-resource instances
Not all instances are resources, so not all InstanceDefinitions found
for a deferred rule will correspond to a Resource. They will lack a
resource type, which is how they are detected and reported as errors.
The isInstance property of CaretValueRule is used to track rules that
should be deferred to potentially resolve as resources.