Contained resources in StructureDefinitions, export primitives with children (#495)
* JSON StructureDefinition includes children of primitives

Children of primitives are added to StructureDefinitions with
property names prefixed with an underscore. Check for such these
properties when converting to JSON.

* Add resources to StructureDefinition

The "contained" element on a StructureDefinition can contain other
Resources. Since this element is on the StructureDefinition itself,
these Resources are added with a CaretValueRule. Since the Resource
to add may be an inline Instance, these rules must be deferred
until after the InstanceExporter has finished. But, any Resource
that can be fished up can be added.

* Add defined resource to StructureDefinition

A resource that is part of SUSHI's FHIR definitions can also be
added to a StructureDefinition's contained resources. For example,
a ValueSet can be contained in this way.

* Log error when named resource is not found

* Detect non-resource instances

Not all instances are resources, so not all InstanceDefinitions found
for a deferred rule will correspond to a Resource. They will lack a
resource type, which is how they are detected and reported as errors.
The isInstance property of CaretValueRule is used to track rules that
should be deferred to potentially resolve as resources.

* Always defer instance-based no-path CaretValueRule

Assert for CaretValueRule checks isInstance property.
Additional cleanup from review feedback.