id is not required in FSHOnly mode
The configured id is used when generating an IG, but is not used
otherwise. Do not require it in FSHOnly mode. The two remaining required
configuration values are:
- canonical: used to set the url on generated resources
- fhirVersion: used to download FHIR resources

The version property is no longer required. It is used to set the
version on generated resources, but this is an optional property.