Add type definition files for generated parser classes (#524)
* Add type definition files for generated parser classes

Antlr generates a number of classes based on the FSH grammar
definition. These classes are produced in javascript, using es5
function-style classes declarations. The typescript compiler will
generate type definitions for these classes, but the generated
definitions contain errors that will appear when another project
specifies fsh-sushi as a dependency. By manually including type
definitions that are error-free, fsh-sushi can safely be used as a
dependency in other projects.

Because the classes most likely to be of use in other project are the
FSH and FHIR types, they are added as exports to index.js. Note that
the change to package.json which specifies index.js allows these
classes to be included more easily, but does mean that fsh-sushi can
no longer be run from its directory with a simple "node ." command.

* Add build script to clean grammar type definitions

The fixGrammarTypes.js script is called as part of the build process.
This removes the need to manually update type definition files when the
grammar is updated.

* Move fixGrammarTypes out of default output

The "build" directory is the default output directory, which is why it
was and once again is listed in .gitignore.

* Convert fixGrammarTypes to ts

This is a typescript project, so the helper script is also written in
typescript. The script is moved to the "dev" directory, since "build" is
the default output folder when running sushi.

* Additional exports

Co-authored-by: Julia Afeltra <jafeltra@mitre.org>