Translate some ANTLR errors to more user friendly messages (#541)
* Translate some ANTLR errors to more user friendly messages

Provide user friendly messages if we detect the following problems:
* missing space around =
* missing space around ->
* missing space after *

This also addresses one issue that can arise when ANTLR hits a parsing error and tries to keep going -- which is that sometimes we'll process a rule that has missing required tokens.  In this case, we want to be a bit more defensive and due a null check before dereferencing the token.

This also adds a global.d.ts to import jest-extended to allow for jest-extended matchers to work properly in VS Code. This has no effect on the distributed module.

* Fix error translation for caret-path assignment rules