Fix the fixedURI in extensions w/ overridden URL
If you override an extension's URL with a caret rule, the extension definition should fix the 'Extension.url' element value to the overridden URL. Prior to this fix, it fixed it to the canonical-based URL instead.  This caused problems if you tried to access and use that extension in a profile.