Fix custom extension urls (#554)
* Test for containing extensions w/ customized urls

This test ensures that if you override an extension's URL with a caret rule, you can still use that extension in a profile.  It is currently broken, but will be fixed in the next commit.

* Fix the fixedURI in extensions w/ overridden URL

If you override an extension's URL with a caret rule, the extension definition should fix the 'Extension.url' element value to the overridden URL. Prior to this fix, it fixed it to the canonical-based URL instead.  This caused problems if you tried to access and use that extension in a profile.