Allow quantities to be fixed to specializations (#552)
* Tests for fixing a Quantity to a Quantity specialization

Authors should be able to fix a Quantity to a Quantity specialization (like Age).  Currently, SUSHI complains because Quantity and Age are not the same type.

* Allow Quantity to be fixed to Quantity specializations

Allow quantities to be fixed to specializations like Age and Duration. Note that we do not do any extra validation (e.g., we don't ensure that Age uses valid units for Age, or that a Count uses the '1' unit) -- we leave that for the IG Publisher to validate.

* Add fixedAge and patternAge to ElementDefinition

This allows the test code to reference them without needing to use ts-ignore.