Fixing complex objects checks for fixed children (#573)
* Fixing complex objects checks for fixed children

When fixing a complex object to an element, children of that element or
of that element's slicedElement may already have fixed values. When this
happens, make sure that the two values do not conflict with one another.

* Inspect arrays inside of complex FHIR values

FHIR values that are complex types, such as CodeableConcepts, may
contain arrays. Flatten these arrays when checking fixed values on child
elements in order to successfully traverse those FHIR values.

* Use path to find child value of complex type