Find slices of choice elements
When a choice element is sliced, the new element is considered to be a
reslice of the choice slice. This does not line up perfectly with the
way PathParts are constructed, which would use the choice slice name
(such as `valueString`) as the PathPart's base. So, if one of the
elements to search through is a slice of a choice element, look for
other slices on that choice element when trying to find the matching
slice.