Retain choice slice min for choice of exactly 1 type
Usually, we make the min of a new type slice 0 because in a type slice, if min isn't 0, then it's not really giving you much of a choice, is it?

But in the case of applying a type slice to a choice that has been constrained to just one element... we should keep the min since that type slice really is the same as the choice it slices (since it is no choice at all anymore).

Fixes #596