Exit with code 0 when there is no FSH
The IG publisher may run SUSHI when no SUSHI project is present. In this
case, show a message indicating that there are no FSH files to read.
This should not be considered an error by the calling process. There may
also be no configuration present when there are no FSH files. This
should also not be considered an error by the calling process.