Exit with code 0 when there is no FSH (#609)
* Exit with code 0 when there is no FSH

The IG publisher may run SUSHI when no SUSHI project is present. In this
case, show a message indicating that there are no FSH files to read.
This should not be considered an error by the calling process. There may
also be no configuration present when there are no FSH files. This
should also not be considered an error by the calling process.

* Defer error message for no configuration

If there is no configuration present, it may also be the case that there
are no FSH files present. The error message for no configuration is only
displayed if there are available FSH files.

* Read config after confirming there are fsh files