Change configuration format to support separate group names and IDs
Previously, custom groups defined in `sushi-config.yaml` used the
same string for both the `id` and `name` properties for a given group.

For example, given the following YAML:

    groups:
      GroupA:
        name: Group A
        description: The Alpha Group
        resources:
        - StructureDefinition/animal-patient
        - StructureDefinition/arm-procedure

Previously, both the `name` and `id` properties in the generated IG JSON
would have been `Group A`. This caused errors like the following when
building the IG with the FHIR publisher:

> ImplementationGuide.definition.resource[0].groupingId
> id value 'Group A' is not valid

This commit changes SUSHI to set `id` to `GroupA` and `name` to
`Group A`. This avoids the error when building the IG.

Groups where `name` is undefined are still supported:

    groups:
      GroupB:
        description: The Beta Group
        resources:
        ...

In this case, both `id` and `name` will be `GroupB`.