Do not apply Must Support to connected slices
An element being Must Support does not automatically mean that its
slices are also Must Support. However, if the child of a sliced element
is Must Support, the corresponding children on the slices are also
Must Support. See
https://confluence.hl7.org/display/FHIR/Guide+to+Designing+Resources for
more details in the section titled "How do I interpret 'Must Support'
with respect to slicing?"