Group properties when exporting Instances
A primitive property on an Instance may have extensions or other nested
attributes on it. This is represented on the Instance with a property
whose key begins with an underscore. For easier reading, when a property
and its underscored version both exist on an Instance, those properties
should be grouped together on the serialized Instance.