Must Support does not automatically transfer from sliced elements to slices (#653)
* Do not apply Must Support to connected slices

An element being Must Support does not automatically mean that its
slices are also Must Support. However, if the child of a sliced element
is Must Support, the corresponding children on the slices are also
Must Support. See
https://confluence.hl7.org/display/FHIR/Guide+to+Designing+Resources for
more details in the section titled "How do I interpret 'Must Support'
with respect to slicing?"

* Delete mustSupport when creating a new slice

By deleting mustSupport before capturing the original state of a new
slice, we can guarantee that if a rule sets mustSupport on this slice,
it will appear in the differential.