Group properties when exporting Instances (#656)
* Group properties when exporting Instances

A primitive property on an Instance may have extensions or other nested
attributes on it. This is represented on the Instance with a property
whose key begins with an underscore. For easier reading, when a property
and its underscored version both exist on an Instance, those properties
should be grouped together on the serialized Instance.

* Reorder instance properties recursively

When reordering instance properties for export to JSON, recursively
traverse the instance and reorder every non-array object. Only the
top-level properties have special keys that have to come first. For all
other objects, follow the rule of "keep a property and its underscore
partner adjacent".

* Move orderedCloneDeep to InstanceDefinition.ts

InstanceDefinition is the only class that uses orderedCloneDeep. If
other classes need it later on, it can move to a shared location at that
time.