Log dependencies individually when extracted from IG (#669)
* Log dependencies individually when extracted from IG

When a configuration is extracted from an ImplementationGuide resource,
info messages are logged for the elements of the configuration. The
dependencies array may have many elements. To avoid an overly-long
message, each element of the array is logged separately.

* Consistent log format for array elements in extracted config

Any array element in a configuration that is extracted from an
ImplementationGuide resource is logged the same way. Currently, these
are the dependencies and fhirVersion elements.