Shrinkwrap (#675)
* Use npm-shrinkwrap instead of package.lock

The only difference between npm-shrinkwrap and package-lock is that
npm-shrinkwrap will be published, but package-lock will not.  Node
discourages use of npm-shrinkwrap in most cases, but specifically
identifies CLI tools that will be globally installed as one VALID
use case.

We need this because we just got bit by an unintended dependency upgrade.
ANTLR 4.9 introduced a dependency on Node 14.

* Only allow patch releases for antlr4

It seems that antlr4 does not follow semantic versioning, as they
introduced a breaking change from 4.8 to 4.9.