Add ParamRuleSet class
A ParamRuleSet is now a distinct entity. RuleSet returns to its original
parameter-less form. A ParamRuleSet has one or more rules, each of which
can be basically anything prior to substitution. Because there are now
multiple rules that start with KW_RULESET, the error case for an empty
RuleSet is slightly different. An error is still emitted, but no RuleSet
is added to the FSHDocument.