Allow more formatting variations in InsertRule with parameters
When writing an InsertRule with parameters, it is no longer required to
use whitespace to separate the RuleSet name and the start of the
parameter list.
The parameter list may contain newlines. This allows authors to write an
InsertRule with one parameter on each line, which is useful if parameter
values are long.
Test cases for inserting parameterized RuleSets have additional checks
for the contents of the generated RuleSets.