Simplify insert rule lexer behavior
Lexer rules that create tokens used for insert rules now use only one
mode instead of 3. The RULESET_REFERENCE token contains the name of the
ruleset to insert and, optionally, the list of parameters to apply to
that ruleset. An additional helper function is used to extract these two
strings from the token.