Space between rule set name and parameter list is optional
When defining a rule set, a space is permitted but not required between
the rule set name and the parameter list. This is consistent with the
way rule sets are referenced by insert rules.

Add explanatory comment to switchToSecretLogger: raising the logging
level means that the logger's format functions are still called, but
nothing is written to the logger's transport. Silencing the logger would
prevent the format functions from being called.