Import Parameterized RuleSet (#687)
* Import a RuleSet with a parameter list

The parameter list must be separated from the RuleSet's name by a space,
because I haven't figured out how to make that better in antlr yet. The
parameters don't do anything yet, but they are present on the RuleSet.

* Allow substitutions in ValueSetRule

The system or strength in a ValueSetRule may be a parameter substitution
instead of a literal value.

* Add ParamRuleSet class

A ParamRuleSet is now a distinct entity. RuleSet returns to its original
parameter-less form. A ParamRuleSet has one or more rules, each of which
can be basically anything prior to substitution. Because there are now
multiple rules that start with KW_RULESET, the error case for an empty
RuleSet is slightly different. An error is still emitted, but no RuleSet
is added to the FSHDocument.

* Add parameters to InsertRule

An inserted RuleSet may also have parameters provided. Store these
parameters in the InsertRule.

The next planned feature is for the RuleSet with the applicable
parameters to be created.

* Apply parameters when inserting RuleSet

When an InsertRule refers to a parameterized RuleSet, apply the provided
parameters to generate the instance of the RuleSet that will be applied
for that rule.

* Do not unescape whitespace when parsing parameters

Only commas and right parentheses should be unescaped when parsing
parameters. This is because they only are escaped in order to correctly
separate parameters and end the parameter list. Escaped whitespace
characters will be unescaped when the generated RuleSet is parsed.

* Apply parameters with a regular expression

Using a regular expression with the global flag allows the replace
function to replace all instances of a parameter.

* Stylistic tidying

* Support nested parameterized RuleSets

If the document used to generate a RuleSet also contains
appliedRuleSets, merge that document's appliedRuleSets into the current
document.

* Simplify grammar for rules in ParamRuleSet

Turns out token negation works just fine with * instead of +. Thanks,
Nick. A side effect of this change is that whitespace is included in the
contents, but that doesn't affect the parsing of the generated RuleSets.

* Allow more formatting variations in InsertRule with parameters

When writing an InsertRule with parameters, it is no longer required to
use whitespace to separate the RuleSet name and the start of the
parameter list.
The parameter list may contain newlines. This allows authors to write an
InsertRule with one parameter on each line, which is useful if parameter
values are long.
Test cases for inserting parameterized RuleSets have additional checks
for the contents of the generated RuleSets.

* Simplify insert rule lexer behavior

Lexer rules that create tokens used for insert rules now use only one
mode instead of 3. The RULESET_REFERENCE token contains the name of the
ruleset to insert and, optionally, the list of parameters to apply to
that ruleset. An additional helper function is used to extract these two
strings from the token.

* Allow multiple uses of the same RuleSet with the same parameters

When a RuleSet is used more than one time with the same parameters, it
is not necessary to generate the applied RuleSet again.

Add params to test for InsertRule.

Change token name in lexer grammar for clarity.

* Store appliedRuleSets by string

A stringified array suffices as a unique key for applied rule sets. The
immutable library is, sadly, not necessary here.

Clarify error message when a parameterized RuleSet is not found.

When parsing a generated RuleSet, guarantee the parent document is
restored by placing it in a finally block.

* Replace parameters simultaneously

All parameters in a parameterized RuleSet have their values substituted
simultaneously. This means that chained replacement no longer occurs.

Add test cases for generated RuleSet with syntax errors, which are
logged in the importer, and nonsensical but syntactically valid rules,
which are not logged by the importer.

* Emit warning for unused parameters in RuleSet

* Define FSHErrorListener subclasses

When importing an actual FSH document, errors should be printed. When
importing a generated FSH document, errors should be collected. Define
FSHErrorPrinter and FSHErrorCollector classes to handle these scenarios.
The abstract parent class FSHErrorListener keeps the logic for building
error messages.

* Collect syntax errors from generated documents

If a syntax error occurs when parsing the document for a generated rule
set, do not emit that error immediately. Instead, collect it using
FSHErrorCollector so that the error can be emitted with source
information for the originating insert rule for that generated rule set.

* Allow whitespace in substitution tokens

* Collect messages more consistently when parsing generated FSH

Errors and warnings that occur when parsing generated FSH may originate
from FSHImporter or from the parser itself. When parsing generated FSH,
temporarily set the log level very high so that console output is
suppressed. Then, once the generated FSH is fully parsed, restore the
logger to normal operation, and emit the collected events as single
messages.

Thanks to @cmoesel for the good plan on this one.

* Do not add undefined to lists of rules

When parsing an insert rule, it is possible that no rule can be created.
In this case, visitInsertRule will return undefined. Do not add this
undefined value to the list of rules for the entity being parsed.

* Space between rule set name and parameter list is optional

When defining a rule set, a space is permitted but not required between
the rule set name and the parameter list. This is consistent with the
way rule sets are referenced by insert rules.

Add explanatory comment to switchToSecretLogger: raising the logging
level means that the logger's format functions are still called, but
nothing is written to the logger's transport. Silencing the logger would
prevent the format functions from being called.