Set source info for generated rule sets
When a rule set is generated by applying parameters, set the source info
for the rule set and the rules it contains based on the source info for
the original parameterized ruleset. This allows error messages in the
exporter to more accurately reference the location of the rules
responsible for the errors.