Redeclaring a contained extension should only be a warning
Since redeclaring a contained (non-inline) extension is a virtual no-op, and since it usually should not lead to problems, demote it from error to warning.

This also helps avoid introducing errors in IGs that previously re-declared extensions on contentrefs (because they had to) but now don't need to since they are automatically carried over.