Carry over extensions when expanding content references (#721)
* Carry over extensions when expanding content references

Fixes #699

* Remove printed log statements from PathUtils tests

* Add test for avoiding recursively nested contentrefs when unfolding

* Add (failing) tests for double slices w/ more slices

If a ContainsRule has more than one item, even if one of the items is a duplicate, the other items should be processed.

* Do best effort to process all items in a ContainsRule

Previously, an error adding an item from a ContainsRule would sometimes result in the remaining items not being processed.  This commit catches errors adding slices more locally so other items are processed.

* Extensions carried over from contentrefs should be in differential

* Update regression to report # of errors

Also changes SUSHI's exit code to be the number of errors.  Prior to this the exit code was only 0 or 1.

* Redeclaring a contained extension should only be a warning

Since redeclaring a contained (non-inline) extension is a virtual no-op, and since it usually should not lead to problems, demote it from error to warning.

This also helps avoid introducing errors in IGs that previously re-declared extensions on contentrefs (because they had to) but now don't need to since they are automatically carried over.