Update min of sliced element and slices when constraining reslices
When constraining the cardinality of a reslice, it may be necessary to
update the minimum cardinality of the parent slice, that is, the slice
being resliced. Check the sliced element and all ancestor slices when
constraining the cardinality, updating other cardinalities as necessary.

The new parentSliceName field on ElementDefinition is used to track
slice lineage. A slice name may legally contain a forward slash, which
is also used as the character to separate the slice name from the
reslice name. Therefore, splitting a slice name on the forward slash
character is not a reliable way of determining the slice name of
ancestor slices.