Do not add duplicates to value set expansions
Before adding a concept or set of concepts to a value set expansion,
check to see if those concepts are already present in the expansion. If
so, do not add them again. The previous implementation would remove the
code before re-adding. However, this implementation is a closer match to
the behavior of the IG publisher.