Check included code system version, formatting
When a value set includes a code system, and the rule specifies a
version, use that version in the compose element. When expanding a value
set, if the compose element specifies a version, only expand if that
version matches the available code system's definition.

Define variables for some of the longer and more intricate conditional
statements in ValueSetExporter.performExpansion.

Break error statement in ValueSetExporter.performExpansion into multiple
lines to avoid potential problems when run from the IG publisher.