Clean up extensions on primitives
When cleaning up properties on a StructureDefinition, properties are
recursively traversed to find and remove those that were added by SUSHI
to help in the export process. Properties that were added by SUSHI start
with an underscore character. Properties that contain extra attributes
of primitives also start with an underscore character, but will have a
sibling property with the same name without the underscore. Therefore,
if such a sibling property is present, the property being examined is a
true member of the StructureDefinition, and should be included in the
cleanup traversal.