Expand ValueSet (#763)
* Expand ValueSet

Authors can request a value set expansion from SUSHI by setting caret
value rules on the value set. Expansion is only possible for specific
concepts and known code systems. If there are filter operators, included
value sets, or unknown code systems involved, expansion will fail.

* Do not add duplicates to value set expansions

Before adding a concept or set of concepts to a value set expansion,
check to see if those concepts are already present in the expansion. If
so, do not add them again. The previous implementation would remove the
code before re-adding. However, this implementation is a closer match to
the behavior of the IG publisher.

* Check included code system version, formatting

When a value set includes a code system, and the rule specifies a
version, use that version in the compose element. When expanding a value
set, if the compose element specifies a version, only expand if that
version matches the available code system's definition.

Define variables for some of the longer and more intricate conditional
statements in ValueSetExporter.performExpansion.

Break error statement in ValueSetExporter.performExpansion into multiple
lines to avoid potential problems when run from the IG publisher.

* Included code system in expansion must be complete or example

* CodeSystems must be content complete in expansion

When expanding a ValueSet, if it contains CodeSystems, those CodeSystems
must be present, and the content attribute on each of those CodeSystems
must be "complete". Otherwise, do not allow expansion.

* Use consistent entity names in warning messages

* Add timestamp and total to ValueSet expansion

FHIR requires a timestamp on ValueSet expansions. It is set to the time
at which the expansion occurred. The optional total field is also
populated, since it is easy enough to calculate.

* Fix comment about required value of CodeSystem.content

* Use ISO date format in ValueSet expansion

Tests are updated to check that the timestamp conforms to the FHIR
dateTime type.