OnlyRule with profile on sliced element uses profiles for intersections (#775)
When an OnlyRule with one or more profiles is applied to a sliced
element, try to use the profile (rather than just the base type) when
finding the type intersection between the new types and the types that
currently exist on a slice. Previously, when the base type was used for
this calculation, a slice which was constrained to a profile would find
an empty type intersection, even when the type being applied to the
sliced element was the same profile as the one on the slice.