Add logical model entity to FSH grammar (#778)
* Add logical model entity to FSH grammar

For details about logical models, visit the official FHIR documentation
http://hl7.org/fhir/R4/structuredefinition.html.

A logical model is a special kind of StructureDefinition that follows
some additional constraints and allows authors to construct more or less
arbitrary data structures within FHIR. As such, any FSH rule that can
appear on a StructureDefinition can also appear on a logical model. A
logical model definition may also contain addElementRules, which allow
authors to add new elements to the logical model. This rule incorporates
cardinality, flags, and types, allowing for concise definitions in many
cases.

* Add Resource to grammar

A Resource is another StructureDefinition variety. It accepts the same
metadata and rules as a logical model.

* Add resource as a possible entity

* Resource contains any number of metadata

Forgot the * on that one. Oops.