Refactor loadExternalDependencies to return Promise<void>
The existing implementation was strange because it returned an array of promises that all resolved to the same FHIRDefinitions instance as the one that was passed in as an argument.  So the actual values were redundant and useless (and in fact were never used).  So what we really need is a single promise that just resolves when everything is loaded.