Skip nonexistent files when traversing directories (#788)
* Skip nonexistent files when traversing directories

A symbolic link with an invalid target will be listed when reading a
directory, but it will not be possible to retrieve information about it.
Therefore, it should be filtered out before continuing to traverse
directories. This is primarily being done to resolve an issue with emacs
lock files.

* Return empty list if stat fails

There are cases where a file may appear to exist according to access
tests, but will not exist according to stat. Catch this case and return
an empty list.

* Unconditionally create and remove test symlink

On some systems, existsSync would return false for the test symlink,
which would result in it still existing after running the test cases.
Unconditionally perform operations with this symlink, and ignore errors
that result from those operations.